Updated inMemory-POS

+++ assignments/js/inMemory-POS/assets/js/CustomerScript.js
@@ -56,14 +56,64 @@ function bindCustomerTrEvents() {
 
 $('#deleteCustomer').click(function () {
     let cus_id = $('#txtCusID').val();
-    deleteCustomer(cus_id);
-    getAllCustomers();
-})
+
+    let consent= confirm("Do you want to delete.?");
+    if (consent) {
+        let response= deleteCustomer(cus_id);
+        if (response){
+            alert("Customer Deleted");
+            getAllCustomers();
+        }else{
+            alert("Customer Not Removed..!");
+        }
+    }
+});
 
 function deleteCustomer(id) {
     for (let i = 0; i < customerDB.length; i++) {
         if (customerDB[i].cusId == id) {
             customerDB.splice(i, 1);
+            return true;
         }
+        return false;
     }
 }
+
+$('#updateCustomerBtn').click(function () {
+    let cus_id = $('#customerId').val();
+    let cus_name = $('#customerName').val();
+    let cus_address = $('#customerAddress').val();
+    let cus_salary = $('#customerSalary').val();
+
+    for (let i = 0; i < customerDB.length; i++) {
+        if (customerDB[i].cusId == cus_id) {
+            customerDB[i].cusName = cus_name;
+            customerDB[i].cusAddress = cus_address;
+            customerDB[i].cusSalary = cus_salary;
+        }
+    }
+    getAllCustomers();
+    bindCustomerTrEvents();
+
+});
+
+$('#btnSearchCustomer').click(function () {
+    let cus_id = $('#txtCusID').val();
+    $("#tblCustomer").empty();
+
+    for (let i = 0; i < customerDB.length; i++) {
+        if (customerDB[i].cusId == cus_id) {
+            let row = `<tr>
+                        <td>${customerDB[i].cusId}</td>
+                        <td>${customerDB[i].cusName}</td>
+                        <td>${customerDB[i].cusAddress}</td>
+                        <td>${customerDB[i].cusSalary}</td>
+                        </tr>`;
+
+            $("#tblCustomer").append(row);
+            bindCustomerTrEvents();
+
+        }
+    }
+
+})
